refactor(web): extract GoBackButton into ssr=false component
Restores the canGoBack conditional (hidden when no history) by moving
the button and handleBack logic into a client-only dynamic import,
avoiding the SSR hydration mismatch.
